상태관리 사용
