+++ package.json
@@ -7,6 +7,7 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
+    "save": "^2.3.3",
     "vue": "^2.6.6",
     "vue-router": "^3.0.1",
     "vuex": "^3.1.0"
