+++ public/index.html
@@ -15,4 +15,5 @@
     <div id="app"></div>
     <!-- built files will be auto injected -->
   </body>
+  <script src="js/load-image.all.min.js"></script>
 </html>
