+++ src/store.js
@@ -1,0 +1,23 @@
+// 상태 관리 파일
+import Vue from 'vue'
+import Vuex from 'vuex'
+
+Vue.use(Vuex);
+
+
+export const store=new Vuex.Store({
+
+  state:{
+    signIn: false
+  },
+  getters:{
+    getSignIn(state){
+      return state.signIn;
+    }
+  },
+  mutations:{
+    signOut(state, payload){
+      state.signIn=false;
+    }
+  },
+});
